refactor: refactor button, move them out of button folder
