--- src/components/ui/buttons/index.ts
@@ -1,16 +1,0 @@
-import ExploreMoreScrollButton from "./ExploreMoreScrollButton";
-import type { ExploreMoreScrollButtonProps } from "./ExploreMoreScrollButton";
-import GetEarlyAccessButton from "./GetEarlyAccessButton";
-import type { GetEarlyAccessButtonProps } from "./GetEarlyAccessButton";
-import JoinDiscordButton from "./JoinDiscordButton";
-import GithubCtaButton from "./GitHubCtaButton";
-import ViewJobsScrollButton from "./ViewJobsScrollButton";
-
-export {
-  ExploreMoreScrollButton,
-  GetEarlyAccessButton,
-  JoinDiscordButton,
-  GithubCtaButton,
-  ViewJobsScrollButton,
-};
-export type { ExploreMoreScrollButtonProps, GetEarlyAccessButtonProps };
