+++ src/components/ui/index.ts
@@ -1,6 +1,28 @@
 import AnnouncementBar from "./AnnouncementBar/AnnouncementBar";
 import Nav from "./Nav";
 import CopyRightText from "./CopyRightText";
-import Footer from "./Footer";
 
-export { AnnouncementBar, CopyRightText, Footer, Nav };
+import ExploreMoreScrollButton from "./ExploreMoreScrollButton";
+import type { ExploreMoreScrollButtonProps } from "./ExploreMoreScrollButton";
+
+import Footer from "./Footer";
+import GetEarlyAccessButton from "./GetEarlyAccessButton";
+import GithubCtaButton from "./GithubCtaButton";
+import JoinDiscordButton from "./JoinDiscordButton";
+
+import ViewJobsScrollButton from "./ViewJobsScrollButton";
+import { ViewJobsScrollButtonProps } from "./ViewJobsScrollButton";
+
+export {
+  AnnouncementBar,
+  CopyRightText,
+  ExploreMoreScrollButton,
+  Footer,
+  Nav,
+  GetEarlyAccessButton,
+  GithubCtaButton,
+  JoinDiscordButton,
+  ViewJobsScrollButton,
+};
+
+export type { ExploreMoreScrollButtonProps, ViewJobsScrollButtonProps };
